Java-Output-Formatter (printf)
Hackerrank exercise to demonstrate use of printf();
Was my first exposure to printf() and after some searching across the net, finally got to grips with how this works.
When run, please enter any String and a number between 0 and 999. This program will format your input, setting the String to left-justified and will pad any 1-2 digit number to ensure 3 digits are printed (e.g. 1 -> 001)